Fix notification sending issue and refactor code
The commit fixes an issue with sending notifications by refactoring the code in the `approve.php` file. The `pageid` and `pagename` variables were updated to default values when necessary. Additionally, a new function `insertNotification` was added to handle inserting notifications into the database and sending FCM messages. The `sendFCMMessage` function was updated to take a `topic` parameter instead of a `to` parameter. The `functions.php` file was also updated to include the new `insertNotification` function.