Update admin/items/update.php with notification feature
This commit adds a notification feature to the admin/items/update.php file. When an item becomes available, users who have requested to be notified about the item will receive a warning notification. The commit also includes deleting all users from the "notifyme" table after sending the notification and updating the "isnotify" field to 0 in the "items" table. The commit message follows the standard format for Git commit messages.